moved create grid index to grid class
